# ecommerce-app/openapi.yaml
openapi: 3.1.0
...
info:
  title: ShopKart E-Commerce Storefront API
  version: 1.0.0
  summary: Customer-facing storefront API for the ShopKart e-commerce application.
  description: |
    REST API powering the ShopKart storefront: authentication, profile & address
    management, product catalogue browsing, cart & coupon handling, checkout /
    orders, and verified product reviews.

    ## Authentication
    Stateless **JWT Bearer** tokens. Obtain a token from `POST /api/auth/signup`
    or `POST /api/auth/login`, then send it as `Authorization: Bearer <token>`.
    Tokens expire after 7 days (configurable via `JWT_EXPIRES_IN`). Logout is
    client-side (drop the token); `POST /api/auth/logout` is a no-op convenience.

    ## Core domain rule
    A purchase is recorded immutably as `user -> order -> order_items`. Each
    `order_item` snapshots the product name, variant name, SKU and unit price
    paid at that time, so order history never changes if a product is later
    renamed, repriced or archived.

    ## Pricing
    - Tax: 8% of `(subtotal - discount)`.
    - Shipping: flat ₹49, waived when `subtotal >= ₹999` (or cart is empty).
    - Coupons: `percent` (optionally capped via `max_discount`) or `fixed`.

    ## Conventions
    - All request and response bodies are `application/json` unless noted.
    - All monetary values are in INR (₹).
    - Timestamps are RFC 3339 / ISO 8601 date-time in UTC.
  contact:
    name: ShopKart API Support
...
  license:
    name: MIT
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/profile:
    get:
      tags: [Profile]
      summary: Get my profile
      operationId: getProfile
      responses:
        '200':
          description: The authenticated user's profile.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Profile]
      summary: Update my profile (name & phone)
      description: Email is immutable and cannot be changed via this endpoint.
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdateRequest' }
            example: { name: Aarav S. Sharma, phone: '+919800000000' }
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ───────────────────────────────  ADDRESSES  ───────────────────────────────
  /api/addresses:
    get:
      tags: [Addresses]
      summary: List my saved addresses
      description: Ordered by default first, then most recently created.
      operationId: listAddresses
      responses:
        '200':
          description: Saved addresses.
          content:
            application/json:
              schema:
                type: object
                required: [addresses]
                properties:
                  addresses:
                    type: array
                    items: { $ref: '#/components/schemas/Address' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Addresses]
      summary: Add a new address
      description: |
        The first address a user creates becomes the default automatically.
        Setting `is_default: true` unsets the default flag on all other addresses.
      operationId: createAddress
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressInput' }
            example:
              label: Home
              recipient_name: Aarav Sharma
              phone: '+919812345678'
              line1: 12 Residency Road
              line2: Apt 4B
              city: Bengaluru
              state: Karnataka
              postal_code: '560025'
              country: India
              is_default: true
      responses:
        '201':
          description: Address created.
          content:
            application/json:
              schema:
                type: object
                required: [address]
                properties:
                  address: { $ref: '#/components/schemas/Address' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/addresses/{id}:
    parameters:
      - $ref: '#/components/parameters/AddressId'
    put:
      tags: [Addresses]
      summary: Update an address
      operationId: updateAddress
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressInput' }
      responses:
        '200':
          description: Address updated.
          content:
            application/json:
              schema:
                type: object
                required: [address]
                properties:
                  address: { $ref: '#/components/schemas/Address' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Addresses]
      summary: Delete an address
      operationId: deleteAddress
      responses:
        '200':
          description: Address deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
              example: { ok: true }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/addresses/{id}/default:
    parameters:
      - $ref: '#/components/parameters/AddressId'
    put:
      tags: [Addresses]
      summary: Set an address as the default
      description: Marks this address as default and clears the flag on all others.
      operationId: setDefaultAddress
      responses:
        '200':
          description: Default updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
              example: { ok: true }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ───────────────────────────────  PRODUCTS  ───────────────────────────────
  /api/products:
    get:
      tags: [Products]
      summary: Browse, search, filter and sort products
      description: |
        Returns active (non-archived) products, each decorated with price range,
        aggregate rating, total stock across variants, and a primary image.
        Supports full-text-ish search (`q`), category and price/rating/stock
        filters, sorting and simple page-based pagination.
      operationId: listProducts
      security: []
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/CategoryFilter'
        - $ref: '#/components/parameters/MinPrice'
        - $ref: '#/components/parameters/MaxPrice'
        - $ref: '#/components/parameters/MinRating'
        - $ref: '#/components/parameters/InStock'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of products.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductListResponse' }
              example:
                products:
                  - id: 1
                    name: Aurora Wireless Headphones
                    slug: aurora-wireless-headphones
                    category: Electronics
                    brand: Aurora
                    description: Over-ear ANC headphones with 40h battery.
                    is_archived: false
                    created_at: '2026-07-01T09:00:00Z'
                    min_price: '4999.00'
                    max_price: '5999.00'
                    total_stock: '37'
                    avg_rating: '4.60'
                    review_count: '18'
                    image: https://cdn.example.com/products/aurora-1.jpg
                page: 1
                limit: 24
...
  /api/cart:
    get:
      tags: [Cart]
      summary: Get my cart (optionally previewing a coupon)
      description: |
        Returns the current cart with line items and computed totals. Passing
        `?coupon=CODE` previews the discount; an invalid code is reported in the
        `coupon_error` field (HTTP still 200) rather than as an error response.
      operationId: getCart
      parameters:
        - $ref: '#/components/parameters/CouponQuery'
      responses:
        '200':
          description: The cart with totals.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
              example:
                cart_id: 7
                items:
                  - cart_item_id: 31
                    quantity: 2
                    variant_id: 108
                    sku: AUR-HP-BLK
                    variant_name: Black
                    unit_price: '4999.00'
                    stock: 12
                    is_active: true
                    product_id: 1
                    product_name: Aurora Wireless Headphones
                    image: https://cdn.example.com/products/aurora-1.jpg
                coupon: { code: WELCOME10, description: 10% off your first order }
                coupon_error: null
                subtotal: '9998.00'
                discount: '999.80'
                tax: '719.86'
                shipping_fee: '0'
                total: '9718.06'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Cart]
      summary: Clear the cart
      operationId: clearCart
      responses:
        '200':
          description: The now-empty cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List my orders
      description: Order history for the authenticated user, newest first.
      operationId: listOrders
      responses:
        '200':
          description: Order history.
          content:
            application/json:
              schema:
                type: object
                required: [orders]
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/OrderListItem' }
              example:
                orders:
                  - id: 1001
                    order_number: SK-2026-001001
                    status: paid
                    payment_status: paid
                    total: '9718.06'
                    placed_at: '2026-07-24T10:35:00Z'
                    item_count: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Orders]
      summary: Place an order from the cart
      description: |
        Creates an order from the current cart. Re-checks stock (locking variant
        rows), snapshots the chosen delivery address and each line item, applies
        an optional coupon, decrements stock and clears the cart — all in one
        transaction. The order starts as `status: placed`, `payment_status: unpaid`.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderCreateRequest' }
            example: { address_id: 15, coupon_code: WELCOME10 }
      responses:
        '201':
          description: Order placed. Returns the new order id and number.
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order: { $ref: '#/components/schemas/OrderRef' }
              example:
                order: { id: 1001, order_number: SK-2026-001001 }
        '400':
          description: Missing address, empty cart, or invalid coupon.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                address: { value: { error: 'Please choose a delivery address.' } }
                empty: { value: { error: 'Your cart is empty.' } }
                coupon: { value: { error: 'Invalid coupon code.' } }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: The delivery address was not found for this user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'Delivery address not found.' }
        '409':
          description: A cart item became unavailable or short on stock during checkout.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                unavailable: { value: { error: '"Aurora Wireless Headphones" is no longer available.' } }
                stock: { value: { error: 'Only 1 of "Aurora Wireless Headphones" left in stock.' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token obtained from `/api/auth/login` or `/api/auth/signup`.
        Send as `Authorization: Bearer <token>`. Payload carries `id`, `email`,
        `role` and `name`; expires after 7 days by default.

  # ─────────────────────────────  PARAMETERS  ─────────────────────────────
  parameters:
    AddressId:
      name: id
      in: path
      required: true
      description: Address identifier owned by the authenticated user.
      schema: { type: integer, format: int32, minimum: 1 }
      example: 15
...
    CouponQuery:
      name: coupon
      in: query
      required: false
      description: Optional coupon code to preview against the cart totals.
      schema: { type: string }
      example: WELCOME10
    SearchQuery:
      name: q
      in: query
      required: false
      description: Case-insensitive keyword matched against product name, description and brand.
      schema: { type: string }
      example: headphones
    CategoryFilter:
      name: category
      in: query
      required: false
      description: Exact category name to filter by (see `/api/products/categories`).
      schema: { type: string }
      example: Electronics
    MinPrice:
      name: min_price
      in: query
      required: false
      description: Minimum product price (INR), inclusive.
      schema: { type: number, minimum: 0 }
      example: 1000
    MaxPrice:
      name: max_price
      in: query
      required: false
      description: Maximum product price (INR), inclusive.
      schema: { type: number, minimum: 0 }
      example: 6000
    MinRating:
      name: min_rating
      in: query
      required: false
      description: Minimum average rating, inclusive.
      schema: { type: number, minimum: 0, maximum: 5 }
      example: 4
    InStock:
      name: in_stock
      in: query
      required: false
      description: When `true`, only products with total stock greater than zero are returned.
      schema: { type: string, enum: ['true', 'false'] }
      example: 'true'
    Sort:
      name: sort
      in: query
      required: false
      description: Sort order. Defaults to `newest` when omitted or unrecognised.
      schema:
        type: string
        enum: [price_asc, price_desc, newest, rating]
        default: newest
      example: rating
    Page:
      name: page
      in: query
      required: false
      description: 1-based page number.
      schema: { type: integer, format: int32, minimum: 1, default: 1 }
      example: 1
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Clamped server-side to the range 1–100.
      schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 24 }
      example: 24

  # ─────────────────────────────  RESPONSES  ─────────────────────────────
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'Name, email, and password are required.' }
    Unauthorized:
      description: Authentication is required or the token is invalid/expired.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            missing: { value: { error: 'Please log in to continue.' } }
            expired: { value: { error: 'Your session has expired. Please log in again.' } }
    NotFound:
      description: The requested resource does not exist (or is not owned by the caller).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'Not found.' }
...
  schemas:
    Error:
      type: object
      description: |
        Standard error body returned by every failing endpoint. The API returns
        a single human-readable `error` message (no code/details envelope).
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
          description: Human-readable, client-safe error message.
          example: 'Please log in to continue.'
    OkResponse:
      type: object
      description: Generic success acknowledgement.
      required: [ok]
      properties:
        ok: { type: boolean, const: true }
...
    DecimalString:
      type: string
      description: |
        A fixed-precision decimal serialised as a string (PostgreSQL NUMERIC).
        Parse before arithmetic. Two decimal places for currency.
      pattern: '^-?\d+(\.\d+)?$'
      example: '4999.00'

    # ── Users & auth ──────────────────────────────────────────────────────
    User:
      type: object
      description: A registered user. The password hash is never returned.
      required: [id, name, email, role]
      properties:
        id: { type: integer, format: int32, readOnly: true, example: 42 }
        name: { type: string, example: Aarav Sharma }
        email: { type: string, format: email, example: aarav@example.com }
        phone:
          type: [string, 'null']
          description: Optional contact phone.
          example: '+919812345678'
        role:
          type: string
          enum: [customer, admin]
          default: customer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2026-07-24T10:30:00Z'
...
    ProfileUpdateRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, example: Aarav S. Sharma }
        phone:
          type: [string, 'null']
          example: '+919800000000'

    # ── Addresses ─────────────────────────────────────────────────────────
    Address:
      type: object
      required: [id, user_id, recipient_name, phone, line1, city, state, postal_code, country, is_default]
      properties:
        id: { type: integer, format: int32, readOnly: true, example: 15 }
        user_id: { type: integer, format: int32, readOnly: true, example: 42 }
        label:
          type: [string, 'null']
          description: Friendly name such as "Home" or "Work".
          example: Home
        recipient_name: { type: string, example: Aarav Sharma }
        phone: { type: string, example: '+919812345678' }
        line1: { type: string, example: 12 Residency Road }
        line2:
          type: [string, 'null']
          example: Apt 4B
        city: { type: string, example: Bengaluru }
        state: { type: string, example: Karnataka }
        postal_code: { type: string, example: '560025' }
        country: { type: string, default: India, example: India }
        is_default: { type: boolean, example: true }
        created_at: { type: string, format: date-time, readOnly: true, example: '2026-07-24T10:30:00Z' }
    AddressInput:
      type: object
      required: [recipient_name, phone, line1, city, state, postal_code]
      additionalProperties: false
      properties:
        label: { type: [string, 'null'], example: Home }
        recipient_name: { type: string, minLength: 1, example: Aarav Sharma }
        phone: { type: string, minLength: 1, example: '+919812345678' }
        line1: { type: string, minLength: 1, example: 12 Residency Road }
        line2: { type: [string, 'null'], example: Apt 4B }
        city: { type: string, minLength: 1, example: Bengaluru }
        state: { type: string, minLength: 1, example: Karnataka }
        postal_code: { type: string, minLength: 1, example: '560025' }
        country: { type: string, default: India, example: India }
        is_default:
          type: boolean
          default: false
          description: Setting `true` clears the default flag on all other addresses.
          example: true

    # ── Products ──────────────────────────────────────────────────────────
...
    ProductListItem:
      type: object
      description: A product decorated with aggregate price/rating/stock and a primary image.
      required: [id, name, category, is_archived, min_price, max_price, total_stock, avg_rating, review_count]
      properties:
        id: { type: integer, format: int32, example: 1 }
        name: { type: string, example: Aurora Wireless Headphones }
        slug: { type: [string, 'null'], example: aurora-wireless-headphones }
        category: { type: string, example: Electronics }
        brand: { type: [string, 'null'], example: Aurora }
        description: { type: [string, 'null'], example: Over-ear ANC headphones with 40h battery. }
        is_archived: { type: boolean, example: false }
        created_at: { type: string, format: date-time, example: '2026-07-01T09:00:00Z' }
        min_price: { $ref: '#/components/schemas/DecimalString' }
        max_price: { $ref: '#/components/schemas/DecimalString' }
        total_stock:
          type: string
          description: Sum of stock across active variants (NUMERIC/bigint as string).
          example: '37'
        avg_rating: { $ref: '#/components/schemas/DecimalString' }
        review_count:
          type: string
          description: Number of reviews (bigint serialised as string).
          example: '18'
        image:
          type: [string, 'null']
          format: uri
          description: URL of the primary product image, or null if none.
          example: https://cdn.example.com/products/aurora-1.jpg
...
    ProductListResponse:
      type: object
      required: [products, page, limit]
      properties:
        products:
          type: array
          items: { $ref: '#/components/schemas/ProductListItem' }
        page: { type: integer, format: int32, description: Echoed current page., example: 1 }
        limit: { type: integer, format: int32, description: Effective (clamped) page size., example: 24 }

    # ── Cart ──────────────────────────────────────────────────────────────
    CartItem:
      type: object
      required: [cart_item_id, quantity, variant_id, sku, variant_name, unit_price, stock, is_active, product_id, product_name]
      properties:
        cart_item_id: { type: integer, format: int32, example: 31 }
        quantity: { type: integer, format: int32, minimum: 1, example: 2 }
        variant_id: { type: integer, format: int32, example: 108 }
        sku: { type: string, example: AUR-HP-BLK }
        variant_name: { type: string, example: Black }
        unit_price: { $ref: '#/components/schemas/DecimalString' }
        stock: { type: integer, format: int32, minimum: 0, example: 12 }
        is_active: { type: boolean, example: true }
        product_id: { type: integer, format: int32, example: 1 }
        product_name: { type: string, example: Aurora Wireless Headphones }
        image:
          type: [string, 'null']
          format: uri
          example: https://cdn.example.com/products/aurora-1.jpg
    CartCoupon:
      type: [object, 'null']
      description: The applied coupon summary, or null if none is applied.
      properties:
        code: { type: string, example: WELCOME10 }
        description: { type: [string, 'null'], example: 10% off your first order }
    Cart:
      type: object
      required: [cart_id, items, coupon, coupon_error, subtotal, discount, tax, shipping_fee, total]
      properties:
        cart_id: { type: integer, format: int32, example: 7 }
        items:
          type: array
          items: { $ref: '#/components/schemas/CartItem' }
        coupon: { $ref: '#/components/schemas/CartCoupon' }
        coupon_error:
          type: [string, 'null']
          description: Populated (with HTTP 200) when a previewed coupon is invalid for this cart.
          example: null
        subtotal: { $ref: '#/components/schemas/DecimalString' }
        discount: { $ref: '#/components/schemas/DecimalString' }
        tax:
          allOf: [{ $ref: '#/components/schemas/DecimalString' }]
          description: 8% of (subtotal − discount).
        shipping_fee:
          allOf: [{ $ref: '#/components/schemas/DecimalString' }]
          description: Flat ₹49, or 0 when subtotal ≥ ₹999 or cart is empty.
        total: { $ref: '#/components/schemas/DecimalString' }
...
    OrderCreateRequest:
      type: object
      required: [address_id]
      additionalProperties: false
      properties:
        address_id:
          type: integer
          format: int32
          minimum: 1
          description: Id of a delivery address owned by the user.
          example: 15
        coupon_code:
          type: [string, 'null']
          description: Optional coupon code to apply to the order.
          example: WELCOME10
    OrderRef:
      type: object
      description: Minimal reference returned when an order is placed.
      required: [id, order_number]
      properties:
        id: { type: integer, format: int32, example: 1001 }
        order_number: { type: string, example: SK-2026-001001 }
    OrderListItem:
      type: object
      required: [id, order_number, status, payment_status, total, placed_at, item_count]
      properties:
        id: { type: integer, format: int32, example: 1001 }
        order_number: { type: string, example: SK-2026-001001 }
        status: { $ref: '#/components/schemas/OrderStatus' }
        payment_status: { $ref: '#/components/schemas/PaymentStatus' }
        total: { $ref: '#/components/schemas/DecimalString' }
        placed_at: { type: string, format: date-time, example: '2026-07-24T10:35:00Z' }
        item_count: { type: integer, format: int32, description: Total quantity across line items., example: 2 }
    OrderStatus:
      type: string
      description: Lifecycle status of an order.
      enum: [placed, paid, packed, shipped, delivered, cancelled]
      example: paid
    PaymentStatus:
      type: string
      enum: [unpaid, paid]
      example: paid
...